Filter repo to only keep the ControllerClientInfo
